自动编译修复proto编译后文件

+++ mxshop-api/goods-web/forms/banner.go
@@ -2,6 +2,6 @@ package forms
 
 type BannerForm struct {
 	Image string `form:"image" json:"image" binding:"url"`
-	Index int    `form:"index" json:"index" binding:"required"`
+	Index uint64 `form:"index" json:"index" binding:"required"`
 	Url   string `form:"url" json:"url" binding:"url"`
 }
